Split SelectionRulesSpecs
